Remove Rust
In the new CI, we will go back to installing the Rust version required
at job time.

This increases a bit the time (~1 minute), and may fail due to network
issues, but we will not be using it to check incoming PRs (so no need to
re-run it like in the past) and we will anyway need to depend on other
resources external to GitHub.

If it fails too many times, then we can look into installing all the
required versions in the image.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>